requirements.txt
fastapi
uvicorn
sqlalchemy
python-multipart
python-jose[cryptography]
passlib[bcrypt]
google-generativeai
python-docx
python-pptx
python-dotenv